requirements.txt
numpy>=1.26
scikit-learn>=1.3
google-cloud-storage>=2.10
google-cloud-bigquery>=3.11
